Authorization requirement on ADD, EDIT, DELETE operations

+++ EndTerm/Controllers/Api/FavouritesController.cs
@@ -4,7 +4,10 @@ using System.Net.Mime;
 using System.Text.Json;
 using System.Threading.Tasks;
 using EndTerm.Models;
+using EndTerm.Models.Request;
 using EndTerm.Repository;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 
@@ -31,32 +34,41 @@ namespace EndTerm.Controllers.Api
             _favouritesItemRepository = favouritesItemRepository;
             _advertisementRepository = advertisementRepository;
         }
-        
-        private Task<IdentityUser> GetCurrentUserAsync() => _userManager.GetUserAsync(HttpContext.User);
-        
+
+        [Authorize(AuthenticationSchemes = 
+            JwtBearerDefaults.AuthenticationScheme)]
         [HttpGet("favourites")]
         public IEnumerable<FavouritesItem> GetFavouritesItems()
         {
+            var email = User.Claims.First().Value;
             var favourites = _favouritesRepository.GetAllFavourites().First(
-                f => f.UserId == GetCurrentUserAsync().Result.Id);
+                f => f.UserId == GetCurrentUser(email).Id);
             var favouriteItems = _favouritesItemRepository.GetAllFavouritesItems().Where(
                 f => f.FavouritesId == favourites.Id);
             return favouriteItems;
 
         }
         
+        /// <summary>
+        /// Add advertisement to favourites
+        /// </summary>
+        /// <param name="request"></param>
+        /// <returns></returns>
+        [Authorize(AuthenticationSchemes = 
+            JwtBearerDefaults.AuthenticationScheme)]
         [HttpPost("favourites/add")]
-        public IActionResult AddFavourite(JsonElement request)
+        public IActionResult AddFavourite(FavouriteItemRequest request)
         {
-            var advertisementId = request.GetProperty("advertisementId").GetInt32();
+            var email = User.Claims.First().Value;
+            var advertisementId = request.AdvertisementId;
             var advertisement = _advertisementRepository.GetAdvertisement(advertisementId);
             var favourites = _favouritesRepository.GetAllFavourites().LastOrDefault(
-                f => f.UserId == GetCurrentUserAsync().Result.Id) ?? new Favourites
+                f => f.UserId == GetCurrentUser(email).Id) ?? new Favourites
             {
-                User = GetCurrentUserAsync().Result,
-                UserId = GetCurrentUserAsync().Result.Id
+                User = GetCurrentUser(email),
+                UserId = GetCurrentUser(email).Id
             };
-
+            
             var favouriteItem = new FavouritesItem
             {
                 Advertisement = advertisement,
@@ -69,12 +81,26 @@ namespace EndTerm.Controllers.Api
             return Ok("Successful");
         }
         
+        [Authorize(AuthenticationSchemes = 
+            JwtBearerDefaults.AuthenticationScheme)]
         [HttpDelete("favourites/{favouritesItemId}")]
         public IActionResult DeleteCategory(int favouritesItemId)
         {
+            var email = User.Claims.First().Value;
+            var user = _userManager.FindByEmailAsync(email).Result;
+            var favourites = _favouritesRepository.GetAllFavourites().Last(
+                f => f.UserId == GetCurrentUser(email).Id);
+            var item = _favouritesItemRepository.GetFavouritesItem(favouritesItemId);
+            if (item.FavouritesId != favourites.Id) return BadRequest("This item does not belong to you");
             var result = _favouritesItemRepository.Delete(favouritesItemId);
             if (result == null) return BadRequest("Item not found");  
             return Ok("Deleted from favourites");
+
+        }
+
+        private IdentityUser GetCurrentUser(string email)
+        {
+            return _userManager.FindByEmailAsync(email).Result;
         }
     }
 }
